service/requirements.txt
asgiref==3.9.1
asttokens==3.0.0
decorator==5.2.1
django==4.2.23
djangorestframework==3.16.1
drf-yasg==1.21.10
exceptiongroup==1.3.0
executing==2.2.0
inflection==0.5.1
ipdb==0.13.13
ipython==8.18.1
jedi==0.19.2
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.51
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.2
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
stack-data==0.6.3
tomli==2.2.1
traitlets==5.14.3
typing-extensions==4.14.1
uritemplate==4.2.0
wcwidth==0.2.13
